Initial port from unittest to pytest of unit/utils/test_gitfs.py to test_gitfs_locks.py
